ingestor/requirements.txt
cfgrib==0.9.12.0
xarray==2024.11.0
rasterio==1.3.11
eccodes==1.6.1
numpy==2.1.3
psycopg2-binary==2.9.10
requests==2.32.3
schedule==1.2.2
python-dotenv==1.0.1
shapely==2.0.6
pyproj==3.7.0
